spike: move non-page file outside `pages/vscode`
nextjs interprets files inside `pages` as actual web pages which
`luxon.ts` is not.
